Refactored and refined analysis
