--- helpers/analysis.py
@@ -1,28 +1,0 @@
-from afinn import Afinn
-from vaderSentiment.vaderSentiment import SentimentIntensityAnalyzer
-import numpy, math
-
-afinn = Afinn(language="en")
-analysis = SentimentIntensityAnalyzer().polarity_scores
-
-def afinn_score(c):
-    word_count = len(c.split())
-    score = ((afinn.score(c) / word_count) / 5)
-    return score
-
-def calc_sentiment(c):
-    vader_sentiment = analysis(c)["pos"] - analysis(c)["neg"]
-    afinn_sentiment = afinn_score(c)
-    return round(((vader_sentiment + afinn_sentiment) / 2), 3)
-
-def weighted_stats(sentiments, likes):
-    weighted_mean = round(numpy.average(sentiments, weights=likes), 3)
-    weighted_std = round(math.sqrt(numpy.average((sentiments-weighted_mean)**2, weights=likes)), 3)
-    return {"weighted_mean": weighted_mean, "weighted_std": weighted_std}
-
-def find_quartiles(list):
-    first_quartile = round(numpy.percentile(list, 25, interpolation = 'midpoint'), 3)
-    third_quartile = round(numpy.percentile(list, 75, interpolation = 'midpoint'), 3)
-    avg_sentiment = round(numpy.median(list), 3)
-    output = {"first_quartile": first_quartile, "third_quartile": third_quartile, "median": avg_sentiment}
-    return output
