+++ index.py
@@ -5,7 +5,8 @@ import config, googleapiclient.discovery, json, os
 
 from comments import get_comments
 from ytcomment import YTComment
-from helpers.analysis import calc_sentiment, find_quartiles
+from helpers.get_sentiment import get_sentiment
+from helpers.get_weighted_stats import get_weighted_stats
 
 
 
@@ -27,13 +28,17 @@ def process_video_id():
     yt_video_id = request.get_json('videoId')
     comment_data = get_comments(yt_video_id) # from line 12, it gets "Bh_uMYaykyQ"
 
+    sentiments = [] # List of each comments' sentiments
+    likes = [] # List of likes per comment
+    comments = [] # List of text value per comment
+
     for c in comment_data["items"]:
         comment = YTComment(c)
-        comment_list.append(comment)
+        sentiments.append(comment.sentiment)
+        likes.append(comment.likes)
+        comments.append(comment.value)
 
-    comment_list.sort(key=lambda c: c.sentiment)
-
-    quartiles = find_quartiles([c.sentiment for c in comment_list])
+    weighted_stats = get_weighted_stats(sentiments, likes) # Weighted mean and weighted standard deviation
 
     return jsonify({'yt_video_id': yt_video_id}), 201
