+++ sample_index.py
@@ -1,31 +1,19 @@
 import json
-import weightedstats as ws
-import weightedcalcs as wc
 from ytcomment import YTComment
-from helpers.analysis import calc_sentiment, weighted_stats
+from helpers.get_sentiment import get_sentiment
+from helpers.get_weighted_stats import get_weighted_stats
 
 with open('./comment_data.json') as f:
     comment_data = json.load(f)
 
-calc = wc.Calculator("likes")
-comment_list = []
+sentiments = [] # List of each comments' sentiments
+likes = [] # List of likes per comment
+comments = [] # List of text value per comment
 
 for c in comment_data["items"]:
     comment = YTComment(c)
-    comment_list.append(comment)
+    sentiments.append(comment.sentiment)
+    likes.append(comment.likes)
+    comments.append(comment.value)
 
-comment_list.sort(key=lambda c: c.sentiment)
-
-sentiment_list = [c.sentiment for c in comment_list]
-likes_list = [c.likes for c in comment_list]
-
-weighted_stats = weighted_stats(sentiment_list, likes_list)
-
-print(weighted_stats)
-print(sentiment_list)
-# print([c.value for c in comment_list])
-
-# weighted_median = ws.numpy_weighted_median(sentiment_list, weights=likes_list)
-# print(weighted_median)
-# quartiles = find_quartiles([c.sentiment for c in comment_list])
-# print(quartiles)
+weighted_stats = get_weighted_stats(sentiments, likes) # Weighted mean and weighted standard deviation
